Refine API of `Region`
